+++ pac4j-jwt/src/main/java/org/pac4j/jwt/config/encryption/SecretEncryptionConfiguration.java
@@ -5,15 +5,10 @@ import com.nimbusds.jose.crypto.AESDecrypter;
 import com.nimbusds.jose.crypto.AESEncrypter;
 import com.nimbusds.jose.crypto.DirectDecrypter;
 import com.nimbusds.jose.crypto.DirectEncrypter;
-import com.nimbusds.jwt.EncryptedJWT;
-import com.nimbusds.jwt.JWT;
-import com.nimbusds.jwt.SignedJWT;
 import org.pac4j.core.exception.TechnicalException;
 import org.pac4j.core.util.CommonHelper;
-import org.pac4j.core.util.InitializableObject;
 
 import java.io.UnsupportedEncodingException;
-import java.text.ParseException;
 
 /**
  * Secret encryption configuration.
@@ -21,17 +16,17 @@ import java.text.ParseException;
  * @author Jerome Leleu
  * @since 1.9.2
  */
-public class SecretEncryptionConfiguration extends InitializableObject implements EncryptionConfiguration {
+public class SecretEncryptionConfiguration extends AbstractEncryptionConfiguration {
 
     private String secret;
 
-    private JWEAlgorithm algorithm = JWEAlgorithm.DIR;
-
-    private EncryptionMethod method = EncryptionMethod.A256GCM;
-
-    public SecretEncryptionConfiguration() {}
+    public SecretEncryptionConfiguration() {
+        algorithm = JWEAlgorithm.DIR;
+        method = EncryptionMethod.A256GCM;
+    }
 
     public SecretEncryptionConfiguration(final String secret) {
+        this();
         this.secret = secret;
     }
 
@@ -57,55 +52,12 @@ public class SecretEncryptionConfiguration extends InitializableObject implement
         CommonHelper.assertNotNull("algorithm", algorithm);
         CommonHelper.assertNotNull("method", method);
 
-
         if (!supports(this.algorithm, this.method)) {
             throw new TechnicalException("Only the direct and AES algorithms are supported with the appropriate encryption method");
         }
     }
 
     @Override
-    public String encrypt(final JWT jwt) {
-        init();
-
-        if (jwt instanceof SignedJWT) {
-            // Create JWE object with signed JWT as payload
-            final JWEObject jweObject = new JWEObject(
-                    new JWEHeader.Builder(this.algorithm, this.method).contentType("JWT").build(),
-                    new Payload((SignedJWT) jwt));
-
-            try {
-                // Perform encryption
-                jweObject.encrypt(buildEncrypter());
-            } catch (final JOSEException e) {
-                throw new TechnicalException(e);
-            }
-
-            // Serialise to JWE compact form
-            return jweObject.serialize();
-        } else {
-            // create header
-            final JWEHeader header = new JWEHeader(this.algorithm, this.method);
-
-            try {
-                // encrypted jwt
-                EncryptedJWT encryptedJwt = new EncryptedJWT(header, jwt.getJWTClaimsSet());
-
-                // Perform encryption
-                encryptedJwt.encrypt(buildEncrypter());
-
-                // serialize
-                return encryptedJwt.serialize();
-            } catch (final JOSEException | ParseException e) {
-                throw new TechnicalException(e);
-            }
-        }
-    }
-
-    /**
-     * Build the appropriate encrypter.
-     *
-     * @return the appropriate encrypter
-     */
     protected JWEEncrypter buildEncrypter() {
         try {
             if (DirectDecrypter.SUPPORTED_ALGORITHMS.contains(algorithm)) {
@@ -119,18 +71,6 @@ public class SecretEncryptionConfiguration extends InitializableObject implement
     }
 
     @Override
-    public void decrypt(final EncryptedJWT encryptedJWT) throws JOSEException {
-        init();
-
-        // decrypt
-        encryptedJWT.decrypt(buildDecrypter());
-    }
-
-    /**
-     * Build the appropriate decrypter.
-     *
-     * @return the appropriate decrypter
-     */
     protected JWEDecrypter buildDecrypter() {
         try {
             if (DirectDecrypter.SUPPORTED_ALGORITHMS.contains(algorithm)) {
@@ -149,22 +89,6 @@ public class SecretEncryptionConfiguration extends InitializableObject implement
 
     public void setSecret(final String secret) {
         this.secret = secret;
-    }
-
-    public JWEAlgorithm getAlgorithm() {
-        return algorithm;
-    }
-
-    public void setAlgorithm(final JWEAlgorithm algorithm) {
-        this.algorithm = algorithm;
-    }
-
-    public EncryptionMethod getMethod() {
-        return method;
-    }
-
-    public void setMethod(final EncryptionMethod method) {
-        this.method = method;
     }
 
     @Override
