+++ pac4j-jwt/src/test/java/org/pac4j/jwt/config/signature/ECSignatureConfigurationTests.java
@@ -9,7 +9,7 @@ import org.junit.Test;
 import org.pac4j.core.exception.TechnicalException;
 import org.pac4j.core.util.TestsConstants;
 import org.pac4j.core.util.TestsHelper;
-import org.pac4j.jwt.config.signature.ECSignatureConfiguration;
+import org.pac4j.jwt.util.JWKHelper;
 
 import java.io.UnsupportedEncodingException;
 import java.security.KeyPair;
@@ -70,7 +70,7 @@ public final class ECSignatureConfigurationTests implements TestsConstants {
     @Test
     public void buildFromJwk() throws UnsupportedEncodingException {
         final String json = new ECKey.Builder(ECKey.Curve.P_256, (ECPublicKey) buildKeyPair().getPublic()).build().toJSONObject().toJSONString();
-        ECSignatureConfiguration.buildFromJwk(json);
+        JWKHelper.buildECKeyPairFromJwk(json);
     }
 
     @Test
