+++ pac4j-jwt/src/test/java/org/pac4j/jwt/config/signature/RSASignatureConfigurationTests.java
@@ -9,7 +9,7 @@ import org.junit.Test;
 import org.pac4j.core.exception.TechnicalException;
 import org.pac4j.core.util.TestsConstants;
 import org.pac4j.core.util.TestsHelper;
-import org.pac4j.jwt.config.signature.RSASignatureConfiguration;
+import org.pac4j.jwt.util.JWKHelper;
 
 import java.io.UnsupportedEncodingException;
 import java.security.KeyPair;
@@ -70,7 +70,7 @@ public final class RSASignatureConfigurationTests implements TestsConstants {
     @Test
     public void buildFromJwk() throws UnsupportedEncodingException {
         final String json = new RSAKey.Builder((RSAPublicKey) buildKeyPair().getPublic()).build().toJSONObject().toJSONString();
-        RSASignatureConfiguration.buildFromJwk(json);
+        JWKHelper.buildRSAKeyPairFromJwk(json);
     }
 
     @Test
